docs: follow steps in web app server repository
This web application has been modified to run on an Apache HTTP
Server as a WSGI Application. Therefore, the previous steps listed
in the 'Usage' section should no longer be followed. Updates to the
code—such as file path changes—will cause the application to not run
properly. Instead, steps listed in the Web Application Server
repository should be ran for this application.